Proper handling of the layout keyword
I originally tried the commit from felixniemeyer's branch, but it
simply disabled mangling following all in/out keywords. This
impacted function parameters that included in/out.

This commit has a more complex change, which tracks the layout
keyword as a token, disabling mangling on global in/out variables,
but not function parameters.